requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
google-generativeai>=0.5.0
gTTS==2.5.1
SpeechRecognition==3.10.3
audio-recorder-streamlit==0.0.10